Refactors Net::HTTP adapter to fix Rubocop ClassLength cop
